Only try to add the symlink if it doesn't already exist.

+++ buildenv.sh
@@ -29,8 +29,11 @@ cd "$DESTINATION"
 #virtualenv -q .
 python3 -m venv .
 
-PYPATH=$(ls -d lib/python* | sort | tail -n 1)
-ln -s "$(basename $PYPATH)" "lib/python"
+if [ ! -e lib/python ]
+then
+    PYPATH=$(ls -d lib/python* | sort | tail -n 1)
+    ln -s "$(basename $PYPATH)" "lib/python"
+fi
 
 . bin/activate
